Expose BitSliceIterator and BitIndexIterator (#1864)
